backend/requirements.txt
flask
openai
python-dotenv
yfinance
pyyaml
